# pkg/protocol/oapi/openapi.yml
openapi: "3.0.1"
info:
  title: Tender Management API
  version: "1.0"
  description: |
    API для управления тендерами и предложениями. 

    Основные функции API включают управление тендерами (создание, изменение, получение списка) и управление предложениями (создание, изменение, получение списка).
servers:
  - url: http://localhost:8080/api
    description: Локальный сервер API
paths:
...
  /tenders:
    get:
      summary: Получение списка тендеров
      description: |
        Список тендеров с возможностью фильтрации по типу услуг.

        Если фильтры не заданы, возвращаются все тендеры.
      security:
        - bearerAuth: []
      operationId: getTenders
      parameters:
        - $ref: "#/components/parameters/paginationLimit"
        - $ref: "#/components/parameters/paginationOffset"
        - name: service_type
          description: |
            Возвращенные тендеры должны соответствовать указанным видам услуг.

            Если список пустой, фильтры не применяются.
          in: query
          schema:
            type: array
            items:
              $ref: "#/components/schemas/tenderServiceType"
            example:
              - Construction
              - Delivery
      responses:
        "200":
          description: Список тендеров, отсортированных по алфавиту по названию.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/tender"
        "400":
          description: Неверный формат запроса или его параметры.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "500":
          description: Сервер не готов обрабатывать запросы, если ответ статусом 500 или любой другой, кроме 200.
  /tenders/new:
    post:
      summary: Создание нового тендера
      description: Создание нового тендера с заданными параметрами.
      security:
        - bearerAuth: []
      operationId: createTender
      requestBody:
        description: Данные нового тендера.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  $ref: "#/components/schemas/tenderName"
                description:
                  $ref: "#/components/schemas/tenderDescription"
                serviceType:
                  $ref: "#/components/schemas/tenderServiceType"
                status:
                  $ref: "#/components/schemas/tenderStatus"
                organizationId:
                  $ref: "#/components/schemas/organizationId"
                creatorUsername:
                  $ref: "#/components/schemas/username"
              required:
                - name
                - description
                - serviceType
                - status
                - organizationId
                - creatorUsername
      responses:
        "200":
          description: Тендер успешно создан. Сервер присваивает уникальный идентификатор и время создания.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tender"
        "400":
          description: Неверный формат запроса или его параметры.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "401":
          description: Пользователь не существует или некорректен.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "500":
          description: Сервер не готов обрабатывать запросы, если ответ статусом 500 или любой другой, кроме 200.
  /tenders/my:
    get:
      summary: Получить тендеры пользователя
      description: |
        Получение списка тендеров текущего пользователя.

        Для удобства использования включена поддержка пагинации.
      security:
        - bearerAuth: []
      operationId: getUserTenders
      parameters:
        - $ref: "#/components/parameters/paginationLimit"
        - $ref: "#/components/parameters/paginationOffset"
        - name: username
          in: query
          schema:
            $ref: "#/components/schemas/username"
      responses:
        "200":
          description: Список тендеров пользователя, отсортированный по алфавиту.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/tender"
        "400":
          description: Неверный формат запроса или его параметры.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "401":
          description: Пользователь не существует или некорректен.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "500":
          description: Сервер не готов обрабатывать запросы, если ответ статусом 500 или любой другой, кроме 200.
  /tenders/{tenderId}/status:
    get:
      summary: Получение текущего статуса тендера
      description: Получить статус тендера по его уникальному идентификатору.
      security:
        - bearerAuth: []
      operationId: getTenderStatus
      parameters:
        - name: tenderId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/tenderId"
        - name: username
          in: query
          schema:
            $ref: "#/components/schemas/username"
      responses:
        "200":
          description: Текущий статус тендера.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tenderStatus"
        "400":
          description: Неверный формат запроса или его параметры.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "401":
          description: Пользователь не существует или некорректен.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "500":
          description: Сервер не готов обрабатывать запросы, если ответ статусом 500 или любой другой, кроме 200.
    put:
      summary: Изменение статуса тендера
      description: Изменить статус тендера по его идентификатору.
      operationId: updateTenderStatus
      security:
        - bearerAuth: []
      parameters:
        - name: tenderId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/tenderId"
        - name: status
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/tenderStatus"
        - name: username
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/username"
      responses:
        "200":
          description: Статус тендера успешно изменен.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tender"
        "400":
          description: Неверный формат запроса или его параметры.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "401":
          description: Пользователь не существует или некорректен.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "403":
          description: Недостаточно прав для выполнения действия.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "404":
          description: Тендер не найден.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "500":
          description: Сервер не готов обрабатывать запросы, если ответ статусом 500 или любой другой, кроме 200.
  /tenders/{tenderId}/edit:
    patch:
      summary: Редактирование тендера
      description: Изменение параметров существующего тендера.
      security:
        - bearerAuth: []
      operationId: editTender
      parameters:
        - name: tenderId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/tenderId"
        - name: username
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/username"
      requestBody:
        description: |
          Перечисление параметров и их новых значений для обновления тендера.

          Если значение не передано, оно останется без изменений.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  $ref: "#/components/schemas/tenderName"
                description:
                  $ref: "#/components/schemas/tenderDescription"
                serviceType:
                  $ref: "#/components/schemas/tenderServiceType"
      responses:
        "200":
          description: Тендер успешно изменен и возвращает обновленную информацию.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tender"
        "400":
          description: Данные неправильно сформированы или не соответствуют требованиям.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "401":
          description: Пользователь не существует или некорректен.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "500":
          description: Сервер не готов обрабатывать запросы, если ответ статусом 500 или любой другой, кроме 200.
  /tenders/{tenderId}/rollback/{version}:
    put:
      summary: Откат версии тендера
      description: Откатить параметры тендера к указанной версии. Это считается новой правкой, поэтому версия инкрементируется.
      operationId: rollbackTender
      security:
        - bearerAuth: []
      parameters:
        - name: tenderId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/tenderId"
        - name: version
          in: path
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
          description: Номер версии, к которой нужно откатить тендер.
        - name: username
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/username"
      responses:
        "200":
          description: Тендер успешно откатан и версия инкрементирована.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tender"
        "400":
          description: Неверный формат запроса или его параметры.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "401":
          description: Пользователь не существует или некорректен.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "403":
          description: Недостаточно прав для выполнения действия.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "404":
          description: Тендер или версия не найдены.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponse"
        "500":
          description: Сервер не готов обрабатывать запросы, если ответ статусом 500 или любой другой, кроме 200.
...
components:
  schemas:
    username:
      type: string
      description: Уникальный slug пользователя.
      example: test_user
    tenderStatus:
      type: string
      description: Статус тендер
      enum:
        - Created
        - Published
        - Closed
    tenderServiceType:
      type: string
      description: Вид услуги, к которой относиться тендер
      enum:
        - Construction
        - Delivery
        - Manufacture
    tenderId:
      type: integer
      format: int32
      description: Уникальный идентификатор тендера, присвоенный сервером.
      example: 1
      minimum: 1
    tenderName:
      type: string
      description: Полное название тендера
      maxLength: 100
    tenderDescription:
      type: string
      description: Описание тендера
      maxLength: 500
    tenderVersion:
      type: integer
      description: Номер версии посел правок
      format: int32
      minimum: 1
      default: 1
    organizationId:
      type: integer
      format: int32
      description: Уникальный идентификатор организации, присвоенный сервером.
      example: 1
      minimum: 1
    tender:
      type: object
      description: Информация о тендере
      properties:
        id:
          $ref: "#/components/schemas/tenderId"
        name:
          $ref: "#/components/schemas/tenderName"
        description:
          $ref: "#/components/schemas/tenderDescription"
        serviceType:
          $ref: "#/components/schemas/tenderServiceType"
        status:
          $ref: "#/components/schemas/tenderStatus"
        organizationId:
          $ref: "#/components/schemas/organizationId"
        verstion:
          $ref: "#/components/schemas/tenderVersion"
        createdAt:
          type: string
          description: |
            Серверная дата и время в момент, когда пользователь отправил тендер на создание.
            Передается в формате RFC3339.
          example: 2006-01-02T15:04:05Z07:00
        
      required:
        - id
        - name
        - description
        - serviceType
        - status
        - verstion
        - createdAt
      example:
        id: 550e8400-e29b-41d4-a716-446655440000
        name: Доставка товары Казань - Москва
        description: Нужно доставить оборудовоние для олимпиады по робототехники
        status: Created
        serviceType: Delivery
        verstion: 1
        createdAt: 2006-01-02T15:04:05Z07:00
...
    errorResponse:
      type: object
      description: Используется для возвращения ошибки пользователю
      properties:
        reason:
          type: string
          description: Описание ошибки в свободной форме
          minLength: 5
      required:
        - reason
      example:
        reason: <объяснение, почему запрос пользователя не может быть обработан>
  parameters:
    paginationLimit:
      in: query
      name: limit
      required: false
      description: |
        Максимальное число возвращаемых объектов. Используется для запросов с пагинацией.

        Сервер должен возвращать максимальное допустимое число объектов.
      schema:
        type: integer
        format: int32
        minimum: 0
        maximum: 50
        default: 5
    paginationOffset:
      in: query
      name: offset
      required: false
      description: |
        Какое количество объектов должно быть пропущено с начала. Используется для запросов с пагинацией.
      schema:
        type: integer
        format: int32
        default: 0
        minimum: 0
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer